Do not generate MBR for fastboot.bin
We do not need a valid MBR for fastboot.bin to boot from USB.  When
there is a extended partition, the whole MBR installation procedure will
be broken.  Since we already have poplar-tools to generate MBR and EBR
even for extended partitions.  We do not want to have l-loader to do it
again in a broken way.  So let's drop it here.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>