Makefile: set RECOVERY=1 by default for now
The RECOVERY build option of arm-trusted-firmware is still on the way to
upstream.  Before it gets merged, arm-trusted-firmware is always a
RECOVERY build.  Let's build l-loader as RECOVERY by default for now to
match arm-trusted-firmware.

The default setting of RECOVERY should be changed back to 0, once
arm-trusted-firmware supports RECOVERY build option.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>